Global ctx passed into the provider via FeatureProvider
